requirements.txt
python-telegram-bot
requests==2.31.0
python-dotenv